requirements.txt
torch
transformers
pyyaml
numpy
scikit-learn
nltk
streamlit
fastapi
uvicorn
requests
